requirements.txt
fastapi
uvicorn
requests
python-dotenv
faiss-cpu
sentence-transformers
ddgs
feedparser
gtts
pydantic
python-multipart
whisper
sounddevice
numpy
scipy
pytest
python-docx
reportlab
